fix: edit card
The previous behaviour makes the list to go empty after saving a card.
It was doing a fetch to the database again, however it was never
successful and the cards of that list in redux store is empty.

This commit changes so that on edit the card store item will be replaced
by new info when saved so that no need to fetch again the data from the
backend.

This commit also finds out an miss that member on a card was never
populated from id so rendering will fail. A temporary fix was to
manually set the member detail with name as id and avatar as default.

Members info should be fetch into store as object, when given an id the
member info is taken through the object.